pyproject.toml
[tool.poetry]
name = "URLs_searcher"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
repository = "https://github.com/RenanCampista/post_link_finder/"
packages = [
    { include = "social_network", from = "src" },
    { include = "client", from = "src", format = "wheel" },
    { include = "utils", from = "src" },
    { include = "search_engines", from = "src" }
]

[tool.poetry.scripts]
searcher = "client.searcher:main"
formatter = "client.formatter:main"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
google-api-python-client = ">=2.129.0"
pandas = ">=2.2.2"
googlesearch-python = ">=1.2.5"
poetry-pyinstaller-plugin = "^1.1.8"  # Mover para a seção de dependências

[tool.pylint.'MAIN']
load-plugins = "pylint.extensions.bad_builtin,pylint.extensions.confusing_elif,pylint.extensions.consider_refactoring_into_while_condition,pylint.extensions.check_elif,pylint.extensions.eq_without_hash,pylint.extensions.no_self_use"
init-hook = "import sys; sys.path.append('src/social_network'); sys.path.append('src/client'); sys.path.append('src/utils')"

[tool.pylint.'MESSAGES CONTROL']
disable = "redefined-outer-name,line-too-long"
